+++ src/HttpClient/ExternalClient.ts
@@ -1,0 +1,25 @@
+import { IOContext } from '../service/typings'
+
+import { IOClient } from './IOClient'
+import { InstanceOptions } from './typings'
+
+/**
+ * Used to perform calls to external endpoints for which you have declared outbound access policies in your manifest.
+ */
+export class ExternalClient extends IOClient {
+  constructor(baseURL: string, context: IOContext, options: InstanceOptions) {
+    const {authToken} = context
+
+    super(
+      context,
+      {
+        ...options,
+        baseURL,
+        headers: {
+          ...options.headers,
+          'Proxy-Authorization': authToken,
+        },
+      }
+    )
+  }
+}
